skills/crypto/claviger/requirements.txt
eciespy>=0.4.1
requests>=2.28.0
